integration test new service, cleaned repo
