+++ backend/old/main.py.old-first
@@ -1,0 +1,67 @@
+from fastapi import FastAPI, HTTPException
+from pydantic import BaseModel
+import json
+from typing import List
+from models import *
+
+app = FastAPI()
+
+tracks = []
+playlists = []
+
+# json files database
+# Load the tracks from the JSON file
+with open("./db/tracks.json", "r") as f:
+    tracks_json = json.load(f)
+
+# Load the playlists from the JSON file
+with open("./db/playlists.json", "r") as f:
+    playlists_json = json.load(f)
+
+for track in tracks_json:
+    tracks.append(track)
+
+# Read a specific track
+@app.get("/tracks/{track_id}")
+def read_track(track_id: int):
+    for track in tracks:
+        if track['id'] == track_id:
+            return track
+    raise HTTPException(status_code=404, detail="Track not found")
+
+# Create a new track
+@app.post("/tracks")
+def create_track(track: Track):
+    if any(x['id'] == track.id for x in tracks):
+        raise HTTPException(status_code=409, detail="Track already exists")
+    tracks.append(track.dict())
+    return {"id": track.id, "name": track.name, "artist": track.artist, "album": track.album, "genre": track.genre, "duration": track.duration}
+
+# Delete a specific track
+@app.delete("/tracks/{track_id}")
+def delete_track(track_id: int):
+    tracks = [track for track in tracks if track['id'] != track_id]
+    return {"message": "Track deleted"}
+
+# Create a new playlist
+@app.post("/playlists")
+def create_playlist(playlist: Playlist):
+    if any(x['id'] == playlist.id for x in playlists):
+        raise HTTPException(status_code=409, detail="Playlist already exists")
+    playlists.append(playlist.dict())
+    return {"id": playlist.id, "name": playlist.name, "tracks": playlist.tracks}
+
+# Read a specific playlist
+@app.get("/playlists/{playlist_id}")
+def read_playlist(playlist_id: int):
+    for playlist in playlists:
+        if playlist['id'] == playlist_id:
+            return {"id": playlist.id, "name": playlist.name, "tracks": playlist.tracks}
+    raise HTTPException(status_code=404, detail="Playlist not found")
+
+# Delete a specific playlist
+@app.delete("/playlists/{playlist_id}")
+def delete_playlist(playlist_id: int):
+    playlists = [
+        playlist for playlist in playlists if playlist['id'] != playlist_id]
+    return {"message": "Playlist deleted"}
