+++ tests/Dockerfile
@@ -1,0 +1,10 @@
+FROM python:3.9
+
+WORKDIR /tests
+
+RUN pip install --no-cache-dir pytest requests
+
+COPY . .
+
+# CMD ["pytest"]
+
